Added the ability to generate binary files using the _BinaryReader:
1) in _FortranRecord:
    a) added the ability to create FortranRecord and put data into it
    b) streamlined get_ and put_ calls to use mostly the same deep function with
       different interfaces
    c) extended to handle both 4- and 8-byte integres

2) in _BinaryReader:
    a) added file mode option to constuctor, default is 'rb'
    b) removed reference to unused floatSize
    c) added method to close file
    d) added methods to write data to file, including a FortranRecord

Also, removed requirement that a FortranRecord has to be some multiple of the
size of an integer.  Some string records are not.

Added FortranRecordError exception class.

Added unit testing for FortranRecord and BinaryReader in test_binaryreader.py.
Reference output for writing a binary file is provided in test_readBR.ref