Updated the interface to binaryreader per pyne mailing list discussion (9/30/11): * all put_<type>() methods will accept either a scalar or an iterable sequence and handle gracefully * all get_<type>() methods will always return an iterable sequence and the user must extract the scalar
This was tested for the binaryreader class itself and passed the tests.

For the CCCC class, the calls to the binaryreader were updated to be consistent with this new interface,
but I was unable to test.  It is also not clear how comprehensive the tests are.  This may have introduced
bugs/flaws into the CCCC reader.