Add util.py and a helper for finding geometry size
For geometries with a box-shaped graveyard shell (e.g. those
produced by mcnp2cad), the helper function find_graveyard_inner_box()
will return the low and high xyz coordinates of the inner surfaces
of the graveyard.  These, in turn, may be assumed to consitute a
tight bounding box around all the space that may be visited by
particles in a monte carlo simulation.

git-svn-id: https://repos.cae.wisc.edu/svn/svalinn/python/trunk/pydagmc@721 3ca0fc55-144d-4cef-8816-fa8a2f6afc42