Added methods to get decay information and completely rewrote the method to get daughters due to neutron reactions. Also created method to convert EAF formatted nuclide strings to zzaaam form.
_get_daughters() now returns a dictionary with daughters as keys and associated
cross sections as values in a NumPy array.

_get_decay() returns a dictionary with decay children as keys and associated
branch ratios as values.

_convert_eaf() currently does not support parsing of the EAF string '0 ..' that
represents products of a fission reaction. It supports excess spaces, and the
suffixes 'G' and 'M#'. Theoretically it will work for any metastable state up
to 9, which is the maximum allowed in zzaaam format.

Some tests for the new methods have been added. Tests to verify correct cross
section data retrieval are necessary, only the daughters of the
_get_daughters() dictionary are currently being checked.

MJK