Major overhaul of build script to support use in other codes
This build script was original designed for DAG-MCNP5, but with this
overhaul has much more modularity making it easier to reuse for other
codes such as FluDAG.  All the code dependent information has been
moved to the top of the file to be included in an otherwise
standardized script, but bulk of which is common to all tools.

As an aside, by using stubs for code dependent steps, this becomes a
good way to build a CGM/MOAB stack.