This commit makes the cmake system backwards compatible for configuring, pulls out some irrelevant stuff, and cleans.
Backwards compatibility is obtained by defining
the new variables,

  Moab_MOAB_LIBRARY Moab_dagmc_LIBRARY
  Geant4_G4geometry_LIBRARY
  Geant4_G4clhep_LIBRARY
  Geant4_G4global_LIBRARY

from the legacy cmake args whenever the new vars
are not defined but the legacy args are.  This
is done at the dagsolid level so that these can
be reused in the src and test subdirs, as they are.

Irrelevant "PROJECT" statements have been removed
from subdirs.  The pthread define is moved up to
the dagsolid level, as it is common to both subdirs.

ln -s changed to ln -sf so that links can be remade.

Library order corrected in test, just in case one
runs into an old linker.

CMAKE_INSTALL_PREFIX is now derived from DAG_SOLID
if configuring the dagsolid level.

Libraries now determined from find_library rather
than hard coding their names.