let read_res add structure even in non-resonance case
The method read_res is often called first. If the LRP flag for
the material is -1, there's no resonance data, and the method did not
change anything.
However other methods expects _read_res to set structur[nuc]['data'],
added function to fill it with a single entry for mat_id and empty values