Ionization Energy Calculator
First run through of a method in response to an issue opened prior regarding added ionization potential calculation functionality. This method will return values corresponding to values of ionization potentials for all electronic states of elements 1-103. We are currently checking numeric values for accuracy, but wished to propose an initial pull request.