Improve regular expressions in endf.pyx
The original regular expression to pick up the contents lines in (MF,MT)=(1,451),
  CONTENTS_R = re.compile(' +\d{1,2} +\d{1,3} +\d{1,10} +')
is not so restrictive enough that the 141st line in n-058_Ce_136.endf of
ENDF/B-VII.1,

  Collective levels used in direct calcualtions                   5825 1451  138
                                                                  5825 1451  139
    Ncoll  Lmax IDef  Kgs  (Def(1,j),j=2,IDef,2)                  5825 1451  140
      12    4    4   0.0  0.170E+00  0.000E+00                    5825 1451  141

is recognized as a contents line and causes an error:

   >>> ds=ENDFDataSource("n-058_Ce_136.endf")
   Traceback (most recent call last):
     File "<stdin>", line 1, in <module>
     File "/home/furutaka/.local/lib/python3.7/site-packages/pyne/xs/data_source.py", line 739, in __init__
       self.library = endf.Library(fh)
     File "endf.pyx", line 80, in pyne.endf.Library.__init__
     File "endf.pyx", line 183, in pyne.endf.Library._read_headers
   ValueError: invalid literal for int() with base 10: '.0  0.170E+'

According to the "ENDF-6 Formats Manual"
  https://www.nndc.bnl.gov/csewg/docs/endf-manual.pdf#page=58
a contents line is one of CONT Records and should consist of the leading 22
whitespaces and following four (at most 11-digit) integers.

Taking the above specification into account, the new regular expression,
  CONTENTS_R = re.compile('^ {22}([ \d]{10}\d){4}')
is more restrictive and do not cause the same kind of errors as above, although
it is NOT PERFECT because it can match the following line (the first two lines are
just for the ease of counting):
(         1          2          3          4          5          6              )
(2345678901234567890123456789012345678901234567890123456789012345678901234567890)
                      1 2 3 4 5 67 8 9 0 1 23 4 5 6 7 89 0 1 2 3 45825 1451  141

The same error occurs with at least the following three ENDF/B-VII.1
files: n-077_Ir_191.endf, n-077_Ir_193.endf, n-081_Tl_205.endf

In the same way, two other regular expressions, FILE1_R and NUMERICAL_DATA_R
were modified.

One of the duplicated SPACE66_R was deleted.