Improve further regular expressions in endf.pyx.
The way (MF,MT)=(1,451) lines are searched for was modified;
by matching a line as a whole (from the beginning to the end
of line).

On the other hand, by scanning a line as a whole, it is NOT
easy to correctly tell what the contents of a line are.
Apparently, the part of a line left to (MAT,MF=1,MT=451) can
be divided into six parts each with 11 characters if the line
is Data or Contents one (CONT Record of ENDF-102).
Therefore, the six strings are tried to match with
  two "E-less" floating numbers and four I11 intergers (Data)
or
  22 spaces followed by four I11 intergers (Contents).
All lines not matched with are treated as those for Document.