Fix issues on newline character(s) in pyne.endf.Library
The length of a line in the ENDF-6 format files is 81 or
82 (including the end-of-line character(s)), depending on
the setting of the universal newline mode in Python;
The size of read(size) only counts the number of "translated"
newline characters.

On the other hand, the offset of seek(offset) counts the
RAW newline characters (i.e. "untranslated").

To fix the issues, the definition of some of the attributes
have been changed:

- self.chars_til_now :
          offset from the top of the file
          (newline characters : untranslated)

- self.line_length :
          length of a line read e.g. through readline()
          (newline characters : after translated)

- self.offset :
          difference between self.line_length and the length
          of a line in the ENDF-6 format file

- start in self._read_headers() :
          the first line number of a MAT

- self.mat_dict[nuc]['mfs'][mf, mt] :
          (offset byte of the (MF,MT) from the beginning of the
	   ENDF-6 format file (newline characters untranslated),
	   the length of the (MF,MT) in the size of read() (newline
	   characters translated))

The use of the attributes were changed correspondingly.