Repair broken test file
The test file for ENDF (sampletape1) had an error in the first line.  It was
missing the last token, making that line too short and also failing to conform
to the ENDF format standard that we are trying to test.

This test file is derived from the official H-1 neutron interaction file.
Upon reviewing the official version of this file, this error was confirmed.

This commit fixes that error.