Don't rely on odd index
This suggestion makes a number of changes.  In general, it is a little
more verbose with variables for long term clarity.  Specifically, it no
longer relies on the counting the entries in the token_list and acting
on the odd ones defined by (i & 1 == 1).  Instead, it pops tokens off
the list and acts on them according to their characteristics.  Some
passes through the while-loop only consume a single token from
the list while others consume 2 tokens from the list.